Move a few structures and enums from encoder to api
